refactor: refactor Router component
